解决权限更新、月末库存问题

+++ ocean-provider/ocean-security/src/main/java/com/fantechs/security/service/impl/SysAuthRoleServiceImpl.java
@@ -29,16 +29,17 @@ public class SysAuthRoleServiceImpl extends BaseService<SysAuthRole> implements 
     public int updateBatch(List<SysAuthRole> list, Byte menuType) {
         List<Long> menuIds = sysMenuInfoMapper.selectMenuId(menuType);
 
-        if (StringUtils.isNotEmpty(menuIds)) {
+        if (StringUtils.isNotEmpty(menuIds) && StringUtils.isNotEmpty(list)) {
             //清除现有角色菜单权限
             Example example = new Example(SysAuthRole.class);
             Example.Criteria criteria = example.createCriteria();
             criteria.andEqualTo("roleId", list.get(0).getRoleId());
             criteria.andIn("menuId", menuIds);
             smtAuthRoleMapper.deleteByExample(example);
+            if (list.get(0).getMenuId() == 0){
+                return 1;
+            }
         }
-
-
         return smtAuthRoleMapper.updateBatch(list);
     }
 }
